Fix PWA types

+++ whtree/modules/publisher/js/pwa/tests.ts
@@ -1,13 +1,13 @@
-/* eslint-disable */
-/// @ts-nocheck -- Bulk rename to enable TypeScript validation
-
 import * as test from "@mod-system/js/wh/testframework";
 import * as maintenance from './internal/maintenance';
 import * as pwadb from '@mod-publisher/js/pwa/internal/pwadb';
+import type { StoreNames } from "idb";
 
-let appname;
+let appname: string | undefined;
 
 export async function deleteDatabase() {
+  if (!appname)
+    throw new Error("Appname not set, call prepare() first");
   return maintenance.deleteDatabase(appname);
 }
 
@@ -15,7 +15,7 @@ export async function unregisterServiceWorkers() {
   return maintenance.unregisterServiceWorkers();
 }
 
-export async function prepare(setappname) {
+export async function prepare(setappname: string) {
   appname = setappname;
   await unregisterServiceWorkers();
 
@@ -26,7 +26,7 @@ export async function prepare(setappname) {
   await maintenance.clearCache(appname);
 }
 
-async function extractIDBTable(database, table) {
+async function extractIDBTable(database: string, table: StoreNames<pwadb.PWADB>) {
   const db = await pwadb.open(database);
   const keys = await db.getAllKeys(table);
   const rows = await Promise.all(keys.map(key => db.get(table, key)));
@@ -34,8 +34,15 @@ async function extractIDBTable(database, table) {
   return rows;
 }
 
-export async function getSWLog() {
-  return await extractIDBTable(appname, 'pwa-swlog');
+interface SWLog { //TODO can PWA define and restrict this ?
+  event: string;
+  url?: string;
+}
+
+export async function getSWLog(): Promise<SWLog[]> {
+  if (!appname)
+    throw new Error("Appname not set, call prepare() first");
+  return await extractIDBTable(appname, 'pwa-swlog') as SWLog[];
 }
 
 export async function touchPage() {
@@ -45,6 +52,6 @@ export async function forceRefresh() {
   return await triggerUpdate('forcerefresh');
 }
 
-async function triggerUpdate(type) {
+async function triggerUpdate(type: "touchpage" | "forcerefresh") {
   return await test.invoke('mod::publisher/lib/internal/pwa/tests.whlib#TriggerUpdate', type, test.getWin().location.href);
 }
